+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+- Method names for open api spec ([#19](https://github.com/scm-manager/scm-jenkins-plugin/pull/19))
+
 ## 2.2.0 - 2020-11-20
 ### Added
 - Send trigger events to the global jenkins instance to trigger updates for namespace navigator items ([#13](https://github.com/scm-manager/scm-jenkins-plugin/pull/13))